--- a/pyproject.toml
+++ b/pyproject.toml
@@ -32,6 +32,7 @@ orbax-checkpoint = "0.5.16"
 flax = "0.8.4"
 uncertainty-toolbox = "^0.1.1"
 e3x = "^1.0.2"
+vesin = "^0.2.0"
 
 [tool.poetry.extras]
 zntrack = ["zntrack"]