I have reworked the PBP dataset to enhance performance.
Previously, the buffer would be filled with batches using multiprocessing, run dry and then be refilled.
Now, there is a dedicated processing thread which uses MP internally.
That means the buffer is now continuously filled without blocking train step compute.

On my home PC I have achieved a 20 % epoch time reduction compared to the cached dataset.
On my work laptop, it's roughly 30 % slower, but still much faster than the old pbp (old: 6.3 s, new: 1.3 s)


Apart from this dataset, there is also the CSVLogger class which uses tensorflow. 
I think if we write our own logger, we could turn tensorflow into an optional dependency.


TODO:

- [x] add vesin to poetry
- [x] swap out matscipy for vesin everywhere
- [x] remove matscipy dependency
- [x] make sure tests are running
- [x] rerun benchmark
- [x] add stress to batch processor


Update:

After rerunning the benchmark on my workstation, the cached dataset is still faster than the PBB, especially in worst case scenarios (same system size, low batch size, lightweight model).
300 atom system, batch size 4, rmax=6

cached: 1.5 s/epoch
pbp new : 2.5 s/epoch
pbp old: 4.7 s/epoch



With BS 8, n_ens = 8, rmax=5
cached: 5.6 s/epoch
pbp: 6.7 s/epoch